remove mocha and chai and use hapi/lab for testing
